rename Test project to CodeExamples & initial rearrange
