+++ CodeExamples/App.config
