+++ CodeExamples/packages.config
